Move old state code to x/evm/types
