+++ x/evm/types/dump.go
@@ -1,10 +1,10 @@
-package state
+package types
 
 import (
 	ethstate "github.com/ethereum/go-ethereum/core/state"
 )
 
-// RawDump returns a raw state drump.
+// RawDump returns a raw state dump.
 //
 // TODO: Implement if we need it, especially for the RPC API.
 func (csdb *CommitStateDB) RawDump() ethstate.Dump {
