+++ x/evm/types/journal.go
@@ -1,4 +1,4 @@
-package state
+package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
