+++ x/evm/types/state_object.go
@@ -1,4 +1,4 @@
-package state
+package types
 
 import (
 	"bytes"
